+++ include/mvh/rtc.h
@@ -1,0 +1,17 @@
+#ifndef MVH_RTC_H
+#define MVH_RTC_H
+
+#include <stdint.h>
+
+typedef struct {
+    uint16_t year;
+    uint8_t month;
+    uint8_t day;
+    uint8_t hour;
+    uint8_t minute;
+    uint8_t second;
+} rtc_time_t;
+
+void rtc_read(rtc_time_t *time);
+
+#endif